+++ app/__tests__/hooks/usePlatforms.test.ts
@@ -1,0 +1,7 @@
+// TODO
+
+describe("usePlatforms", () => {
+  it("test", () => {
+    expect(true).toBe(true);
+  });
+});
