+++ app/components/ExpiredStampsPanel.tsx
@@ -1,8 +1,9 @@
 import { PlatformSpec } from "@gitcoin/passport-platforms";
 import React, { useContext } from "react";
-import { usePlatforms } from "../config/platforms";
+import { usePlatforms } from "../hooks/usePlatforms";
 import { CeramicContext } from "../context/ceramicContext";
 import InitiateReverifyStampsButton from "./InitiateReverifyStampsButton";
+import { PLATFORM_ID } from "@gitcoin/passport-types";
 
 type StampsListProps = {
   className?: string;
@@ -16,8 +17,8 @@ const ExpiredStampsList = ({ className }: StampsListProps) => {
     <div className={`flex flex-col items-center ${className}`}>
       <div className={`flex flex-wrap justify-center gap-8`}>
         {expiredPlatforms
-          ? Object.values(expiredPlatforms)
-              .map((platform) => getPlatformSpec(platform.platform.platformId))
+          ? Object.keys(expiredPlatforms)
+              .map((platformId) => getPlatformSpec(platformId as PLATFORM_ID))
               .filter((platformSpec): platformSpec is PlatformSpec => !!platformSpec)
               .map((platformSpec) => {
                 // check if platform has onchain providers
